Move util.py into utils.py
